feat: add global exception handling and validation for Category APIs

+++ category-service/src/main/java/com/getir/categoryservice/service/CategoryManager.java
@@ -3,6 +3,7 @@ package com.getir.categoryservice.service;
 import com.getir.categoryservice.dto.CategoryRequest;
 import com.getir.categoryservice.dto.CategoryResponse;
 import com.getir.categoryservice.entity.Category;
+import com.getir.categoryservice.exceptions.ApiException;
 import com.getir.categoryservice.mapper.CategoryMapper;
 import com.getir.categoryservice.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
@@ -32,6 +33,9 @@ public class CategoryManager implements CategoryService {
     @Override
     public CategoryResponse getCategoryBySlugTr(String slugTr) {
         Category category = categoryRepository.findBySlugTr(slugTr);
+        if (category == null) {
+            throw new ApiException("Category not found with slugEn: " + slugTr, HttpStatus.NOT_FOUND);
+        }
         return categoryMapper.toResponse(category);
     }
 
@@ -39,7 +43,7 @@ public class CategoryManager implements CategoryService {
     public CategoryResponse getCategoryBySlugEn(String slugEn) {
         Category category = categoryRepository.findBySlugEn(slugEn);
         if (category == null) {
-            throw new RuntimeException("Category not found with slugEn: " + slugEn);
+            throw new ApiException("Category not found with slugEn: " + slugEn, HttpStatus.NOT_FOUND);
         }
         return categoryMapper.toResponse(category);
     }
@@ -47,6 +51,13 @@ public class CategoryManager implements CategoryService {
 
     @Override
     public CategoryResponse createCategory(CategoryRequest request) {
+        boolean existsTr = categoryRepository.findBySlugTr(request.getSlugTr()) != null;
+        boolean existsEn = categoryRepository.findBySlugEn(request.getSlugEn()) != null;
+
+        if (existsTr || existsEn) {
+            throw new ApiException("Category with same slugTr or slugEn already exists", HttpStatus.BAD_REQUEST);
+        }
+
         Category category = categoryMapper.toEntity(request);
         Category saved = categoryRepository.save(category);
         return categoryMapper.toResponse(saved);
@@ -55,13 +66,18 @@ public class CategoryManager implements CategoryService {
     @Override
     public CategoryResponse updateCategory(UUID id, CategoryRequest request) {
         Category category = categoryRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Category not found"));
+                .orElseThrow(() -> new ApiException("Category not found with id: " + id, HttpStatus.NOT_FOUND));
         categoryMapper.updateEntity(category, request);
         return categoryMapper.toResponse(categoryRepository.save(category));
     }
 
     @Override
     public void deleteCategory(UUID id) {
+        boolean exists = categoryRepository.existsById(id);
+        if (!exists) {
+            throw new ApiException("Category not found with id: " + id, HttpStatus.NOT_FOUND);
+        }
         categoryRepository.deleteById(id);
     }
+
 }
